Move many input enums to their own file
